refactor: changes callnumbers to callnumber
Also starts catalog query in callnumber section
